+++ TombEngine/Objects/TR5/Trap/tr5_ventilator.cpp
@@ -13,6 +13,8 @@ namespace TEN::Entities::Traps
 {
 	void VentilatorEffect(GameBoundingBox* bounds, int intensity, short rot, int speed)
 	{
+		constexpr auto DUST_SIZE_MAX = 64.0f;
+
 		int x, y, z;
 
 		if (abs(intensity) == 1)
@@ -52,13 +54,15 @@ namespace TEN::Entities::Traps
 		auto& part = *GetFreeParticle();
 
 		part.on = 1;
+		part.SpriteSeqID = ID_DEFAULT_SPRITES;
+		part.SpriteID = SPRITE_TYPES::SPR_UNDERWATERDUST;
 		part.sR = 0;
 		part.sG = 0;
 		part.sB = 0;
-		part.dR = part.dG = (48 * speed) / 128;
+		part.dR = part.dG = (48 * speed) / 64;
 		part.colFadeSpeed = 4;
 		part.fadeToBlack = 8;
-		part.dB = (speed * ((GetRandomControl() & 8) + 48)) / 128;
+		part.dB = (speed * ((GetRandomControl() & 8) + 48)) / 64;
 		part.blendMode = BlendMode::Additive;
 		part.life = part.sLife = (GetRandomControl() & 3) + 20;
 
@@ -132,7 +136,8 @@ namespace TEN::Entities::Traps
 		part.zVel = (speed * part.zVel) / 128;
 		part.maxYvel = 0;
 		part.gravity = 0;
-		part.flags = SP_NONE;
+		part.flags = SP_DEF | SP_EXPDEF;
+		part.sSize = part.size = Random::GenerateFloat(DUST_SIZE_MAX / 2, DUST_SIZE_MAX);
 	}
 
 	void InitializeVentilator(short itemNumber)