+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The dates are in European standard format where date is presented as **YYYY-MM-DD**.
 TombEngine releases are located in this repository (alongside with Tomb Editor): https://github.com/TombEngine/TombEditorReleases
 
-## [Version 1.8.2]
+## [Version 1.8.2](link to release) - yyyy-MM-DD
 
 ### Bug fixes
 * Fixed crashes when shooting, if gunflash or gunshell objects are not present in a level.
@@ -11,9 +11,12 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 * Fixed Wraith objects not working correctly in flipped rooms.
 * Fixed caustics not rendered correctly if texture compression was enabled.
 * Fixed exclusion blend mode not working correctly.
+* Updated MakeExplosion and EmitBlood to spawn correct effects when used underwater.
 
 ### Lua API changes
 * Fixed `Moveable.GetJointPosition` not returning correct results if moveable is invisible or not rendered.
+* Added `EmitWaterfallMist` under Effects.
+* Added `EmitSink` under Effects.
 
 ## [Version 1.8.1](https://github.com/TombEngine/TombEditorReleases/releases/tag/v1.8.1) - 2025-03-29